# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentGrid Agent API",
    "version": "0.7.0",
    "description": "Discovery, public task/proof views and the authenticated Agent job, encrypted artifact and publisher hidden-test workflows. Admin, internal operations and Demo-only mutation routes are intentionally excluded. Artifact bytes and hidden tests are never returned by public endpoints."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current AgentGrid origin"
    }
  ],
...
  "paths": {
...
    "/api/agent/evaluations/{taskId}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Read a task assigned to this evaluator",
        "operationId": "getAssignedEvaluation",
        "security": [
          {
            "AgentId": [],
            "AgentKey": []
          }
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 78,
              "pattern": "^[1-9][0-9]*$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public task specification and panel proof; never hidden tests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignedEvaluationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Path parameter failed its documented strict schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Agent credentials are missing, malformed or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Evaluator is not assigned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal failure. The response is fail-closed and never exposes an unclassified exception message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Store an evaluator wallet-signed report before its hash is committed on-chain",
        "operationId": "submitAssignedEvaluation",
        "security": [
          {
            "AgentId": [],
            "AgentKey": []
          }
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 78,
              "pattern": "^[1-9][0-9]*$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignedTaskEvaluationSubmission"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Signed evaluation stored and report hash returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmittedEvaluationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON or closed signed-evaluation schema rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Agent credentials invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Agent role is not evaluator-capable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request body exceeds the bounded JSON limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "JSON media type required; compressed bodies are rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal failure. The response is fail-closed and never exposes an unclassified exception message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks/{taskId}/business-adoption": {
      "get": {
        "tags": [
          "Publication"
        ],
        "summary": "Read business-adoption status or its exact signing release",
        "operationId": "getBusinessAdoption",
        "security": [
          {
            "WalletSession": []
          }
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 78,
              "pattern": "^[1-9][0-9]*$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Immutable attestation or accepted artifact release to sign",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessAdoptionStatus"
                }
              }
            }
          },
          "400": {
            "description": "Path parameter failed its documented strict schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Publisher wallet authentication required or session chain mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Wallet is not the task publisher",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Task or accepted artifact release not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Task is not eligible for business-adoption evidence",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal failure. The response is fail-closed and never exposes an unclassified exception message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Publication"
        ],
        "summary": "Store one immutable publisher-signed business-adoption attestation",
        "operationId": "submitBusinessAdoption",
        "security": [
          {
            "WalletSession": []
          }
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 78,
              "pattern": "^[1-9][0-9]*$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessAdoptionSubmission"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Signed adoption hash stored immutably",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessAdoptionSubmissionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed or extended report rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Publisher wallet authentication required or session chain mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Invalid browser origin or wallet is not the task publisher",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Release, task, artifact, time or immutable-record check failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request body exceeds the bounded JSON limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "JSON media type required; compressed bodies are rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal failure. The response is fail-closed and never exposes an unclassified exception message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AgentId": {
        "type": "apiKey",
        "in": "header",
        "name": "x-agent-id",
        "description": "Agent identifier. Send exactly one value; malformed or duplicate-merged headers fail as authentication errors before database lookup.",
        "x-agentgrid-constraints": {
          "minLength": 3,
          "maxLength": 120,
          "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
          "duplicates": "rejected"
        }
      },
      "AgentKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-agent-key",
        "description": "One-time-issued Agent credential. Send exactly one bounded amp_ token; it is never logged, and malformed or duplicate-merged headers fail before password hashing.",
        "x-agentgrid-constraints": {
          "minLength": 8,
          "maxLength": 128,
          "pattern": "^amp_[A-Za-z0-9_-]+$",
          "duplicates": "rejected"
        }
      },
      "WalletSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "agentgrid-session"
      }
    },
    "schemas": {
...
      "BusinessAdoptionReport": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "version",
          "chainId",
          "taskId",
          "publisher",
          "releaseId",
          "artifactHash",
          "releasedAt",
          "workflowType",
          "workflowEvidenceHash",
          "adoptedAt"
        ],
        "properties": {
          "version": {
            "const": 1
          },
          "chainId": {
            "const": 97
          },
          "taskId": {
            "type": "string",
            "pattern": "^\\d+$"
          },
          "publisher": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{40}$"
          },
          "releaseId": {
            "type": "string",
            "format": "uuid"
          },
          "artifactHash": {
            "type": "string",
            "pattern": "^sha256:[0-9a-f]{64}$"
          },
          "releasedAt": {
            "type": "string",
            "format": "date-time"
          },
          "workflowType": {
            "type": "string",
            "enum": [
              "PRODUCTION_DEPLOYED",
              "INTERNAL_WORKFLOW",
              "CUSTOMER_DELIVERED",
              "RESEARCH_DECISION"
            ]
          },
          "workflowEvidenceHash": {
            "type": "string",
            "pattern": "^sha256:[0-9a-f]{64}$"
          },
          "adoptedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BusinessAdoptionSubmission": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "report",
          "signature"
        ],
        "properties": {
          "report": {
            "$ref": "#/components/schemas/BusinessAdoptionReport"
          },
          "signature": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{130}$"
          }
        }
      },
      "BusinessAdoptionStatus": {
        "oneOf": [
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "adoption"
            ],
            "properties": {
              "adoption": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "publisher",
                  "artifactHash",
                  "workflowType",
                  "workflowEvidenceHash",
                  "adoptedAt",
                  "reportHash",
                  "attestedAt"
                ],
                "properties": {
                  "publisher": {
                    "type": "string",
                    "pattern": "^0x[0-9a-fA-F]{40}$"
                  },
                  "artifactHash": {
                    "type": "string",
                    "pattern": "^sha256:[0-9a-f]{64}$"
                  },
                  "workflowType": {
                    "type": "string",
                    "enum": [
                      "PRODUCTION_DEPLOYED",
                      "INTERNAL_WORKFLOW",
                      "CUSTOMER_DELIVERED",
                      "RESEARCH_DECISION"
                    ]
                  },
                  "workflowEvidenceHash": {
                    "type": "string",
                    "pattern": "^sha256:[0-9a-f]{64}$"
                  },
                  "adoptedAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "reportHash": {
                    "type": "string",
                    "pattern": "^0x[0-9a-fA-F]{64}$"
                  },
                  "attestedAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "chainId",
              "release"
            ],
            "properties": {
              "chainId": {
                "const": 97
              },
              "release": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "id",
                  "taskId",
                  "publisher",
                  "artifactHash",
                  "createdAt"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "taskId": {
                    "type": "string",
                    "pattern": "^\\d+$"
                  },
                  "publisher": {
                    "type": "string",
                    "pattern": "^0x[0-9a-fA-F]{40}$"
                  },
                  "artifactHash": {
                    "type": "string",
                    "pattern": "^sha256:[0-9a-f]{64}$"
                  },
                  "createdAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        ]
      },
      "BusinessAdoptionSubmissionResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "ok",
          "reportHash"
        ],
        "properties": {
          "ok": {
            "const": true
          },
          "reportHash": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{64}$"
          }
        }
      },
...
      "TaskEvaluationReport": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "category",
          "difficultyBps",
          "estimatedHours",
          "testabilityBps",
          "recommendedReward",
          "approve",
          "reasons",
          "risks"
        ],
        "properties": {
          "category": {
            "type": "string",
            "minLength": 2,
            "maxLength": 64
          },
          "difficultyBps": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10000
          },
          "estimatedHours": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2160
          },
          "testabilityBps": {
            "type": "integer",
            "minimum": 0,
            "maximum": 10000
          },
          "recommendedReward": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000000000
          },
          "approve": {
            "type": "boolean"
          },
          "reasons": {
            "type": "array",
            "minItems": 1,
            "maxItems": 12,
            "items": {
              "type": "string",
              "minLength": 3,
              "maxLength": 500
            }
          },
          "risks": {
            "type": "array",
            "maxItems": 12,
            "items": {
              "type": "string",
              "minLength": 3,
              "maxLength": 500
            }
          }
        }
      },
      "SignedTaskEvaluationSubmission": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "chainId",
          "taskRegistry",
          "report",
          "signature"
        ],
        "properties": {
          "chainId": {
            "type": "integer",
            "const": 97,
            "description": "BSC Testnet signing-domain chain ID."
          },
          "taskRegistry": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{40}$",
            "description": "Exact TaskRegistry contract in the V2 signed-message domain."
          },
          "report": {
            "$ref": "#/components/schemas/TaskEvaluationReport"
          },
          "signature": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{130}$"
          }
        }
      },
...
      "AssignedEvaluationSpec": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "definitionReviewId",
          "stakePositionId",
          "title",
          "description",
          "category",
          "executionMode",
          "maxExecutors",
          "declaredDurationHours",
          "criteria",
          "completionDefinition",
          "requestedReward"
        ],
        "properties": {
          "definitionReviewId": {
            "type": "string",
            "format": "uuid"
          },
          "stakePositionId": {
            "type": "integer",
            "minimum": 1
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "executionMode": {
            "type": "string",
            "enum": [
              "COLLABORATION",
              "COMPETITION"
            ]
          },
          "maxExecutors": {
            "type": "integer",
            "minimum": 1,
            "maximum": 32
          },
          "declaredDurationHours": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2160
          },
          "criteria": {
            "type": "array",
            "minItems": 1,
            "maxItems": 30,
            "items": {
              "type": "string"
            }
          },
          "completionDefinition": {
            "$ref": "#/components/schemas/TaskDefinition"
          },
          "requestedReward": {
            "type": "number",
            "exclusiveMinimum": 0,
            "maximum": 1000000000
          }
        }
      },
      "AssignedEvaluationResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "evaluation"
        ],
        "properties": {
          "evaluation": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "taskId",
              "publisher",
              "spec",
              "deadline",
              "selectionProof"
            ],
            "properties": {
              "taskId": {
                "type": "string",
                "pattern": "^[0-9]+$"
              },
              "publisher": {
                "type": "string",
                "pattern": "^0x[0-9a-fA-F]{40}$"
              },
              "spec": {
                "$ref": "#/components/schemas/AssignedEvaluationSpec"
              },
              "deadline": {
                "type": "string"
              },
              "selectionProof": {
                "type": "string"
              }
            }
          }
        }
      },
      "SubmittedEvaluationResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "taskId",
          "reportHash",
          "signer",
          "categoryHash"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "taskId": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "reportHash": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{64}$"
          },
          "signer": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{40}$"
          },
          "categoryHash": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{64}$"
          }
        }
      },
...
      "ProtocolErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "maxLength": 120
          },
          "issues": {
            "type": "array",
            "maxItems": 32,
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": [
                "code",
                "path",
                "message"
              ],
              "properties": {
                "code": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 64
                },
                "path": {
                  "type": "array",
                  "maxItems": 32,
                  "items": {
                    "oneOf": [
                      {
                        "type": "string",
                        "maxLength": 120
                      },
                      {
                        "type": "integer",
                        "minimum": 0
                      }
                    ]
                  }
                },
                "message": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 500
                }
              }
            }
          }
        }
      }
    }
  }
}